+++ Models/Employee.cs
@@ -14,7 +14,7 @@ namespace ERPProject.Models
         public string Surname { get; set; }
         public DateTime Hired { get; set; }
         public DateTime Departed { get; set; }
-        public float Salary { get; set; }
+        public double Salary { get; set; }
         public string Position { get; set; }
     }
 }
